app/backend/requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Pydantic (pin stable version to avoid Rust build issues)
pydantic==2.8.2
pydantic-settings==2.2.1

# Environment
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
#asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
aiomysql==0.2.0
pymysql==1.1.0
greenlet==3.0.3

# Security
python-jose[cryptography]==3.3.0
cryptography==42.0.5

# AI / External APIs
openai==1.30.1
sse-starlette==1.6.5

# Payments
stripe==12.6.0

